Rephrase DiscoveredCatalog as a view model
Its primary responsibility was to format data into something appropriate
for a response. In my book, this qualifies as a view concern.
