+++ crates/control/src/services/mod.rs
@@ -1,7 +1,6 @@
 pub mod builder;
 pub mod builds_root;
 pub mod connectors;
-pub mod discovery;
 pub mod sessions;
 pub mod signatures;
 pub mod subprocess;
